Re-organise pytest-benchmark
